requirements.txt
google-genai
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
websockets
python-dotenv